+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a simple profiler to the probe-rs cli toolkit (#1628)
 - Added MSP432E4 target (MSP432E401Y and MSP432E411Y). (#1139)
 - probe-rs-cli: added `attach` subcommand. (#1672, #1616)
+- Added arm interface support to ftdi probes.
 
 ### Changed